Development

+++ app/Http/Controllers/C_project.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\M_project;
+use App\Models\M_layanan;
 use App\Models\M_team;
 use Illuminate\Support\Facades\Storage;
 
@@ -18,16 +19,18 @@ class C_project extends Controller
      */
     public function index(){
         $project = M_project::latest()->get();
+
         if (request('search')){
             $project->where('nama','like','%'.request('search').'%');
         }
+
         $no = 1;
         $noDetail = 0;
         return view('adminView.project.S_project',[
-            'no'=>$no,
-            'title'=>'project',
-            'project'=>$project,
-            'noDetail'=>$noDetail
+            'no' => $no,
+            'title' => 'Project',
+            'project' => $project,
+            'noDetail' => $noDetail
         ]);
     }
 
@@ -47,7 +50,8 @@ class C_project extends Controller
         }
 
         $project = new M_project;
-        $project->nama=$request->name;
+        $project->nama_project=$request->nama_project;
+        $project->nama_client=$request->nama_client;
         $project->deskripsi=$request->deskripsi;
         $project->jenis=$request->jenis;
         $project->foto_transaksi=$validasi['foto_transaksi'];
@@ -56,15 +60,16 @@ class C_project extends Controller
         $project->keterangan=$request->keterangan;
         $project['pekerja']=json_encode($request->pekerja);
         $project->save();
-        return redirect()->route('R_project.index');
 
+        return redirect("/project")->with("success", "New Project has beed added");
     }
 
     public function addProject(){
-        $pekerja = M_team::latest();
+
         return view('adminView.project.A_project',[
-            'title'=>'addproject',
-            'pekerja'=>$pekerja->get(),
+            "title" => "Add Project",
+            "pekerja" => M_team::latest()->get(),
+            "layanan" => M_layanan::latest()->get()
         ]);
     }
 
@@ -80,6 +85,18 @@ class C_project extends Controller
             'arrayPekerja'=>$arrayPekerja
         ]);
     }
+
+    public function edit(M_project $project){
+        $arrayPekerja = json_decode($project->pekerja);
+        
+        return view('adminView.project.E_project',[
+            'title' => 'Edit Project',
+            'pekerja' => M_team::latest()->get(),
+            'project' => $project,
+            'arrayPekerja' => $arrayPekerja,
+            "layanan" => M_layanan::latest()->get()
+        ]);
+    }
     /**
      * Update the specified resource in storage.
      *
@@ -87,42 +104,51 @@ class C_project extends Controller
      * @param  \App\M_project  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request,$id){
+    public function update(Request $request, M_project $project){
         $validasi = $request->validate([
-            'foto_transaksi' => 'file|image|mimes:jpg,png,jpeg,gif,svg|max:2048',
-            'foto_completed' => 'file|image|mimes:jpg,png,jpeg,gif,svg|max:2048',
+            "nama_client" => "required",
+            "nama_project" => "required",
+            "deskripsi" => "required",
+            "keterangan" => "required",
+            "deadline" => "required",
+            "status" => "required",
+            "jenis" => "required",
+            "foto_transaksi" => "file|image|mimes:jpg,png,jpeg,gif,svg|max:2048",
+            "foto_completed" => "file|image|mimes:jpg,png,jpeg,gif,svg|max:2048"
         ]);
-        $project = M_project::findOrFail($id);
+
         if($request->file('foto_transaksi')){
             if($request->oldfoto_transaksi){
-                Storage::delete('public/'.$request->oldfoto_transaksi);
+                Storage::delete($project->foto_transaksi);
             }
-            $validasi['foto_transaksi']=$request->file('foto_transaksi')->store('imgTransaksi','public');
-        }else {
-            $validasi['foto_transaksi']=$project->foto_transaksi;
+
+            $validasi['foto_transaksi'] = $request->file('foto_transaksi')->store('imgTransaksi');
         }
-        
-        if ($request->file('foto_completed')){
+
+        if($request->file('foto_completed')){
             if($request->oldfoto_completed){
-                Storage::delete('public/'.$request->oldfoto_completed);
+                Storage::delete($project->foto_completed);
             }
-            $validasi['foto_completed']=$request->file('foto_completed')->store('imgComplete','public');
-        }else {
-            $validasi['foto_completed']=$project->foto_completed;
+
+            $validasi['foto_completed'] = $request->file('foto_completed')->store('imgCompleted');
         }
         
-        $project->nama=$request->name;
-        $project->deskripsi=$request->deskripsi;
-        $project->jenis=$request->jenis;
-        $project->foto_transaksi=$validasi['foto_transaksi'];
-        $project->deadline=$request->deadline;
-        $project->status=$request->status;
-        $project->keterangan=$request->keterangan;
-        $project->time_completed=$request->time_completed;
-        $project->foto_completed=$validasi['foto_completed'];
-        $project['pekerja']=json_encode($request->pekerja);
-        $project->update();
-        return redirect()->route('R_project.index');
+        
+
+        
+        // if ($request->file('foto_completed')){
+        //     if($request->oldfoto_completed){
+            //         Storage::delete('public/'.$request->oldfoto_completed);
+        //     }
+        //     $validasi['foto_completed']=$request->file('foto_completed')->store('imgComplete','public');
+        // }else {
+        //     $validasi['foto_completed']=$project->foto_completed;
+        // }
+        
+        $validasi['pekerja'] = json_encode($request->pekerja);
+        $project->update($validasi);
+        
+        return redirect("/project")->with("success", "Project has beed updated");
     }
 
     /**
@@ -133,9 +159,11 @@ class C_project extends Controller
      */
     public function destroy($id){
         $project = M_project::find($id);
-        Storage::delete('public/'.$project->foto_transaksi);
+        if ($project->foto_transaksi) {
+            Storage::delete('public/'.$project->foto_transaksi);
+        }
         Storage::delete('public/'.$project->foto_completed);
         $project->delete();
-        return redirect()->route('R_project.index')->with('sukses');
+        return redirect("/project");
     }
 }
